% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_C_mat}
\alias{create_C_mat}
\title{Creates C matrix for use in mvnorm_prob_each_best}
\usage{
create_C_mat(K)
}
\arguments{
\item{K}{The dimension of C}
}
\value{
A matrix
}
\description{
Creates C matrix for use in mvnorm_prob_each_best
}
