% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_with_control.R
\name{estimate_lm_model}
\alias{estimate_lm_model}
\title{Estimate linear regression model using variational approximation.}
\usage{
estimate_lm_model(
  dat,
  prior = c(int_prior_mean = 40, int_prior_sd = 5, b_prior_sd = 5, df = 3, scale = 5),
  ctr = contr.treatment
)
}
\arguments{
\item{dat}{A dataset}

\item{prior}{A vector giving prior parameters:
intercept prior mean, intercept prior sd, effect prior sd, variance prior df, variance prior scale.
Note that the assumed prior on the variance is Half-t(df, scale) and on effects is Normal(0, effect_sd).}

\item{ctr}{A contrast specification}
}
\description{
Estimate linear regression model using variational approximation.
}
