% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R
\name{estimate_lmm_model}
\alias{estimate_lmm_model}
\title{Estimate the model}
\usage{
estimate_lmm_model(dat, zero_sum = F)
}
\arguments{
\item{dat}{The data (as a data.table)}

\item{zero_sum}{Use sum-to-zero coding}
}
\value{
A list giving the mean and variance of the variational approximation to the fixed effects
}
\description{
Estimate the model
}
