% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_model2.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate linear regression model}
\usage{
estimate_model(
  X,
  y,
  m0 = rep(0, ncol(X)),
  s0 = rep(10, ncol(X)),
  a0 = 0.01,
  b0 = 0.01,
  ...
)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response vector}

\item{m0}{Prior mean}

\item{s0}{Prior SD}

\item{a0}{Prior shape}

\item{b0}{Prior shape}

\item{...}{Other arguments to `vb_lm`}
}
\description{
Estimate linear regression model
}
