% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_model2.R
\name{generate_preference_data}
\alias{generate_preference_data}
\title{Generate potential outcomes if a participant were to receive each treatment}
\usage{
generate_preference_data(
  n = 320,
  m = matrix(c(-3, -1, -1, -1, -2, 2, 0, 0, -2, 0, 2, 0, -1, 1, 1, 3), 4, 4),
  s = 10,
  p = c(0.2, 0.7, 0.05, 0.05)
)
}
\arguments{
\item{n}{The number of participants}

\item{m}{A matrix giving the mean response on each treatment (cols) by preference (rows)}

\item{s}{The variability of response assumed constant across preferences/treatments}

\item{p}{Distribution of preferences in the population}
}
\description{
Generate potential outcomes if a participant were to receive each treatment
}
