% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R
\name{make_design}
\alias{make_design}
\title{Create the design matrix}
\usage{
make_design(zero_sum = FALSE)
}
\arguments{
\item{zero_sum}{Use sum-to-zero permutation invariant coding}
}
\value{
The design matrix
}
\description{
Create the design matrix
}
