% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mass_weighted_urn_design}
\alias{mass_weighted_urn_design}
\title{Mass-weighted urn randomisation
Zhao W. (2015).
Mass weighted urn design--A new randomization algorithm for unequal allocations.
Contemporary clinical trials, 43, 209–216.
https://doi.org/10.1016/j.cct.2015.06.008}
\usage{
mass_weighted_urn_design(target_alloc, sample_size, alpha = 4)
}
\arguments{
\item{target_alloc}{The target allocation ratios}

\item{sample_size}{The number of allocations to generate}

\item{alpha}{Parameter to control imbalance between arms}
}
\value{
A list detailing the mass-weighted-urn process.
}
\description{
Mass-weighted urn randomisation
Zhao W. (2015).
Mass weighted urn design--A new randomization algorithm for unequal allocations.
Contemporary clinical trials, 43, 209–216.
https://doi.org/10.1016/j.cct.2015.06.008
}
