% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R,
%   R/simulate_trial_model2.R
\name{prob_supr}
\alias{prob_supr}
\title{Probability each column is maximum}
\usage{
prob_supr(M)

prob_supr(M)
}
\arguments{
\item{M}{A matrix of Monte Carlo draws}
}
\description{
Probability each column is maximum

Probability each column is maximum
}
