% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R
\name{simulate_accrual_data}
\alias{simulate_accrual_data}
\title{Simulate accrual data}
\usage{
simulate_accrual_data(nsubj = 200, accrual = function(n) 1:n)
}
\arguments{
\item{nsubj}{The number of subjects}

\item{accrual}{An accrual function which returns `nsubj` ordered randomisation times in weeks}
}
\value{
A data.table giving the accrual data
}
\description{
Simulate accrual data
}
