% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R
\name{simulate_general_outcome_data}
\alias{simulate_general_outcome_data}
\title{Simulate outcome data}
\usage{
simulate_general_outcome_data(
  nsims = 50,
  nsubj = 400,
  means = matrix(0, 4, 4, dimnames = list(t = 0:3, a = 0:3)),
  covar = {     R <- matrix(0.5, 4, 4)     diag(R) <- 1     5^2 * R },
  ...
)
}
\arguments{
\item{nsims}{Number of sets of outcome data to simulate}

\item{nsubj}{Number of subjects}

\item{means}{The mean at each time point in each arm}

\item{covar}{The co-variance structure between time points (assumed equal across all arms)}
}
\value{
A data.table giving outcome data
}
\description{
Simulate outcome data
}
