% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R
\name{simulate_longitudinal_trial}
\alias{simulate_longitudinal_trial}
\title{Simulate a longitudinal trial}
\usage{
simulate_longitudinal_trial(
  n_seq = c(100, 150, 200, 250, 400),
  dat = simulate_general_outcome_data(nsims = 1, nsubj = max(n_seq)),
  eff_eps = 0.95,
  sup_eps = 0.95,
  alloc = rep(1/4, 4),
  brar = TRUE,
  make_dt = TRUE,
  zero_sum = FALSE,
  scale = FALSE,
  ...
)
}
\arguments{
\item{n_seq}{The sequence of analyses by number of subjects with 12 week observation}

\item{dat}{The dataset for this particular simulation}

\item{eff_eps}{Effective decision threshold}

\item{sup_eps}{Superiority decision threshold}

\item{alloc}{Initial allocation ratio}

\item{brar}{Update allocation ratio using BRAR}
}
\value{
A list (or data.table) giving the trial results
}
\description{
Simulate a longitudinal trial
}
