% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R
\name{simulate_outcome_data}
\alias{simulate_outcome_data}
\title{Simulate outcome data}
\usage{
simulate_outcome_data(
  nsims = 50,
  nsubj = 200,
  accrual = function(n) 1:n,
  beta = c(0, 2, 4, 6, 0, -2.5, -2.5, -2.5, 0, 2.5, 2.5, 2.5, 0, 0, 0, 0),
  sigma_e = 5,
  sigma_u = 1
)
}
\arguments{
\item{nsims}{Number of sets of outcome data to simulate}

\item{nsubj}{Number of subjects}

\item{accrual}{Accrual function for use in `simulate_accrual_data`}

\item{beta}{Fixed effects parameters}

\item{sigma_e}{Variance of residuals for outcome}

\item{sigma_u}{Variance of subject specific intercepts}
}
\value{
A data.table giving outcome data
}
\description{
Simulate outcome data
}
