% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_model2.R
\name{simulate_preference_trial}
\alias{simulate_preference_trial}
\title{Simulate trial accounting for preference}
\usage{
simulate_preference_trial(
  dat = generate_preference_data(),
  n_seq = c(100, 150, 200, 250, nrow(dat$y)),
  alloc = matrix(1/ncol(dat$y), length(dat$p), ncol(dat$y)),
  sup_eps = 0.95,
  eff_eps = 0.95,
  brar = FALSE,
  min_decision = 20,
  ...
)
}
\arguments{
\item{dat}{Data generated from `generate_preference_data`}

\item{n_seq}{Sequence of interim analyses}

\item{alloc}{Initial allocation probabilities (matrix)}

\item{brar}{Use BRAR?}

\item{min_decision}{The minimum sample size in a subgroup to make decision/switch on BRAR}

\item{...}{Other arguments to `vb_lm` functions}
}
\description{
Simulate trial accounting for preference
}
