% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_model2.R
\name{simulate_standard_trial}
\alias{simulate_standard_trial}
\title{Simulate trial without any knowledge of preference}
\usage{
simulate_standard_trial(
  dat = generate_preference_data(),
  n_seq = c(100, 150, 200, 250, nrow(dat$y)),
  alloc = rep(1/ncol(dat$y), ncol(dat$y)),
  sup_eps = 0.95,
  eff_eps = 0.95,
  brar = FALSE,
  ...
)
}
\arguments{
\item{dat}{Data generated from `generate_preference_data`}

\item{n_seq}{Sequence of interim analyses}

\item{alloc}{Initial allocation probabilities}

\item{brar}{Use BRAR?}

\item{...}{Other arguments to `vb_lm` functions}
}
\description{
Simulate trial without any knowledge of preference
}
