% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tnvar}
\alias{tnvar}
\title{Calculate variance of truncated normal}
\usage{
tnvar(a, b, mu, sigma)
}
\arguments{
\item{a}{Lower bound}

\item{b}{Upper bound}

\item{mu}{Mean of normal}

\item{sigma}{Standard deviation of normal}
}
\value{
Variance of truncated normal
}
\description{
Calculate variance of truncated normal
}
