% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R
\name{trial_as_dt}
\alias{trial_as_dt}
\title{Convert trial list into trial data.table}
\usage{
trial_as_dt(res)
}
\arguments{
\item{res}{A list which is the result of simulate_longitudinal_trial}
}
\value{
The list converted to a data.table with one row being an analysis by treatment result
}
\description{
Convert trial list into trial data.table
}
