% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_model2.R
\name{update_model}
\alias{update_model}
\title{Update linear regression model}
\usage{
update_model(fit, X, y, ...)
}
\arguments{
\item{fit}{A previous `vb_lm` fit}

\item{X}{New design matrix data}

\item{y}{New response vector data}

\item{...}{Other arguments to `update_vb_lm`}
}
\description{
Update linear regression model
}
