% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial_longitudinal.R
\name{simulate_accrual_data}
\alias{simulate_accrual_data}
\title{Simulate accrual data}
\usage{
simulate_accrual_data(
  nsubj = 200,
  accrual = function(n) round(cumsum(rexp(n, 3))),
  dropout = function(n) sample(c(1, Inf), n, replace = T, prob = c(0.2, 0.8))
)
}
\arguments{
\item{nsubj}{The number of subjects}

\item{accrual}{An accrual function which returns `nsubj` ordered randomisation times in weeks}

\item{dropout}{A drop-out function which returns `nsubj` times-to-drop-out, one for each subject.
`Inf` means no drop-out (not implemented)}
}
\value{
A data.table giving the accrual data
}
\description{
Simulate accrual data
}
